a tool for ui and server for getting datanames taking into account the datasets_selected vector

a tool for ui and server for getting datanames taking into account the datasets_selected vector

get_datanames_selected(datasets, datasets_selected)

Arguments

datasets

teal datasets object

datasets_selected

(character) a character vector that says which datasets should be shown and in what order. Names in a vector have to correspond with datasets names.

Value

(character) a character vector