Scatterplot and Regression Model

Scatterplot and Regression Model

tm_g_regression(label = "Regression Analysis", dataname, regressor,
  response, plot_height = c(600, 200, 2000), pre_output = NULL,
  post_output = NULL)

Arguments

label
label shown in the navigation for the item

dataname
name of datasets used to generate the regression plot (just used for labeling)

regressor
(list) Output of teal.devel::data_extract_spec to define the regressor variable from an incoming dataset with filtering and selecting.

response
(list) Output of teal.devel::data_extract_spec to define the response variable from an incoming dataset with filtering and selecting.

plot_height
(numeric) a vector of length three with c(value, min and max) for a slider encoding the plot height.

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

Examples

library(random.cdisc.data)
library(tern)

ASL <- radsl(seed = 1)
keys(ASL) <- c("STUDYID", "USUBJID")

asl_extracted <- data_extract_spec(
  dataname = "ASL",
  columns = columns_spec(
    choices = c("SEX", "AGE", "BMRKR1", "BMRKR2"),
    selected = c("AGE"),
    multiple = TRUE,
    fixed = FALSE
  )
)

app <- init(
  data = cdisc_data(
    ASL = ASL,
    code = 'ASL <- radsl(seed = 1); keys(ASL) <- c("STUDYID", "USUBJID")',
    check = FALSE),
  modules = root_modules(
    tm_g_regression(
      label = "Regression",
      dataname = c("ASL"),
      response = list(asl_extracted),
      regressor = list(asl_extracted)
    )
  )
)

#> |apply filter for ASL
#> |** UI id for modules is teal_modules.root
#> |UI id for module is teal_modules_root.Regression

# NOT RUN {
shinyApp(app$ui, app$server)
# }