Response Plots

Response Plots

tm_g_response(
  label = "Response Plot",
  response,
  x,
  row_facet = NULL,
  col_facet = NULL,
  coord_flip = FALSE,
  count_labels = TRUE,
  rotate_xaxis_labels = FALSE,
  freq = FALSE,
  plot_height = c(600, 400, 5000),
  plot_width = NULL,
  ggtheme = gg_themes,
  ggplot2_args = teal.devel::ggplot2_args(),
  pre_output = NULL,
  post_output = NULL
)

Arguments

label

(character) Label shown in the navigation item for the module.

response

(data_extract_spec or list of multiple data_extract_spec) Which variable to use as the response. You can define one fixed column by using the setting fixed = TRUE inside the select_spec. data_extract_spec must not allow multiple selection in this case.

x

(data_extract_spec or list of multiple data_extract_spec) Which variable to use on the X-axis of the response plot. Allow the user to select multiple columns from the data allowed in teal. data_extract_spec must not allow multiple selection in this case.

row_facet

optional, (data_extract_spec or list of multiple data_extract_spec) Which data columns to use for faceting rows.

col_facet

optional, (data_extract_spec or list of multiple data_extract_spec) Which data to use for faceting columns.

coord_flip

optional, (logical) Whether to flip coordinates between x and response.

count_labels

optional, (logical) Whether to show count labels. Defaults to TRUE.

rotate_xaxis_labels

optional, (logical) Whether to rotate plot X axis labels. Does not rotate by default (FALSE).

freq

optional, (logical) Whether to display frequency (TRUE) or density (FALSE). Defaults to density (FALSE).

plot_height

optional, (numeric) A vector of length three with c(value, min and max) for a slider encoding the plot height.

plot_width

optional, (numeric) A vector of length three with c(value, min and max) for a slider encoding the plot width.

ggtheme

optional, (character) ggplot Theme to be used by default. gg_themes is defined internally as c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic", "void", "test") All themes can be chosen by the user. Defaults to gray.

ggplot2_args

(`ggplot2_args`) object created by [teal.devel::ggplot2_args()] with settings for the module plot. For more details see the help vignette: `vignette("Custom ggplot2_args arguments module", package = "teal.devel")` The argument is merged with options variable `teal.ggplot2_args` and default module setup.

pre_output

optional, shiny.tag with text placed before the output to put the output into context. For example a title.

post_output

optional, shiny.tag with text placed after the output to put the output into context. For example the helpText elements are useful.

Note

For more examples, please see the vignette "Using response plot" via vignette("using-response-plot", package = "teal.modules.general").

Examples

# Response plot with selected response (BMRKR1) and selected x variable (RACE)
library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"),
    check = TRUE
  ),
  modules = root_modules(
    tm_g_response(
      label = "Response Plots",
      response = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(ADSL, c("BMRKR2", "COUNTRY")),
          selected = "BMRKR2",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(ADSL, c("SEX", "RACE")),
          selected = "RACE",
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
#> [INFO] 2022-01-13 20:25:50.1423 pid:1794 token: teal.modules.general Initializing tm_g_response
if (FALSE) {
shinyApp(app$ui, app$server)
}