Create a simple scatterplot

Create a plot with the ggplot2[geom_point] function

tm_g_scatterplot(
  label = "Scatterplot",
  x,
  y,
  color_by = NULL,
  size_by = NULL,
  row_facet = NULL,
  col_facet = NULL,
  plot_height = c(600, 200, 2000),
  plot_width = NULL,
  alpha = c(1, 0, 1),
  shape = shape_names,
  size = c(5, 1, 15),
  max_deg = 5L,
  rotate_xaxis_labels = FALSE,
  ggtheme = gg_themes,
  pre_output = NULL,
  post_output = NULL,
  table_dec = 4
)

Arguments

label
(character) Label shown in the navigation item for the module.

x
(data_extract_spec or list of multiple data_extract_spec) Variable names selected to plot along the x-axis by default.

y
(data_extract_spec or list of multiple data_extract_spec) Variable names selected to plot along the y-axis by default.

color_by
optional (data_extract_spec or list of multiple data_extract_spec) Defines the color encoding. If NULL then no color encoding option will be displayed.

size_by
optional (data_extract_spec or list of multiple data_extract_spec) Defines the point size encoding. If NULL then no size encoding option will be displayed.

row_facet
optional, (data_extract_spec or list of multiple data_extract_spec) Which data columns to use for faceting rows.

col_facet
optional, (data_extract_spec or list of multiple data_extract_spec) Which data to use for faceting columns.

plot_height
optional, (numeric) A vector of length three with c(value, min and max) for a slider encoding the plot height.

plot_width
optional, (numeric) A vector of length three with c(value, min and max) for a slider encoding the plot width.

alpha
optional, (numeric) If scalar then the plot points will have a fixed opacity. If a slider should be presented to adjust the plot point opacity dynamically then it can be a vector of length three with c(value, min, max).

shape
optional, (character) A character vector with the English names of the shape, e.g. c("triangle", "square", "circle"). It defaults to shape_names. This is a complete list from vignette("ggplot2-specs", package="ggplot2").

size
optional, (numeric) If scalar then the plot point sizes will have a fixed size If a slider should be presented to adjust the plot point sizes dynamically then it can be a vector of length three with c(value, min, max).

max_deg
optional, (integer) The maximum degree for the polynomial trend line. Must not be less than 1.

rotate_xaxis_labels
optional, (logical) Whether to rotate plot X axis labels. Does not rotate by default (FALSE).

ggtheme
optional, (character) ggplot Theme to be used by default. gg_themes is defined internally as c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic", "void", "test") All themes can be chosen by the user. Defaults to gray.

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example a title.

post_output
optional, shiny.tag with text placed after the output to put the output into context. For example the helpText elements are useful.

table_dec
optional, (integer) Number of decimal places used to round numeric values in the table.

Note

For more examples, please see the vignette "Using scatterplot" via vignette("using-scatterplot", package = "teal.modules.general").

Examples

# Scatterplot of variables from ADSL dataset
library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"),
    check = TRUE
  ),
  modules = root_modules(
    tm_g_scatterplot(
      label = "Scatterplot Choices",
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(ADSL, c("AGE", "BMRKR1", "BMRKR2")),
          selected = "AGE",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      y = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(ADSL, c("AGE", "BMRKR1", "BMRKR2")),
          selected = "BMRKR1",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      color_by = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(ADSL, c("AGE", "BMRKR1", "BMRKR2", "RACE", "REGION1")),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      size_by = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(ADSL, c("AGE", "BMRKR1", "BMRKR2", "RACE", "REGION1")),
          selected = "AGE",
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      row_facet = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(ADSL, c("AGE", "BMRKR1", "BMRKR2", "RACE", "REGION1")),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      ),
      col_facet = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(ADSL, c("AGE", "BMRKR1", "BMRKR2", "RACE", "REGION1")),
          selected = NULL,
          multiple = FALSE,
          fixed = FALSE
        )
      )
    )
  )
)
#> 2021-12-08 14:46:49.2089 pid:1150 teal.modules.general fun:tm_g_scatterplot [INFO] Initializing tm_g_scatterplot
if (FALSE) {
shinyApp(app$ui, app$server)
}