Create a simple cross-table

Create a table with the table[base] function

tm_table(label, dataname, xvar, yvar, useNA = c("ifany", "no", "always"),
  pre_output = NULL, post_output = NULL)

Arguments

label
(chracter) menu label

dataname
(chracter) name of dataset used to generate table

xvar
(list of data_extract_spec) Specification how the user can select data to get encoded in the rows of the cross table. Please just use single selections inside the columns_spec.

yvar
(list of data_extract_spec) Specification how the user can select data to get encoded in the columns of the cross table. Please just use single selections inside the columns_spec.

useNA
(character) optional pre-selected option indicating how to utilize NA in table display. One of 'ifany', 'always', 'no'. If missing then 'ifany' will be used. If vector then only the first one will be used.

pre_output
(shiny.tag) html tags appended below the output

post_output
(shiny.tag) html tags appended after the output

Examples

library(random.cdisc.data)
library(tern)

ASL <- radsl(seed = 1)
keys(ASL) <- c("USUBJID", "STUDYID")

asl_extract_xvar <- data_extract_spec(
    "ASL",
    columns = columns_spec(
        choices = base::setdiff(names(ASL), keys(ASL)), # strict call of setdiff
        selected = names(ASL)[5],
        multiple = FALSE,
        fixed = FALSE
    )
)

asl_extract_yvar <- data_extract_spec(
    "ASL",
    columns = columns_spec(
        choices = base::setdiff(names(ASL), keys(ASL)), # strict call of setdiff
        selected = names(ASL)[6],
        multiple = FALSE,
        fixed = FALSE
    )
)

app <- init(
  data = cdisc_data(
    ASL = ASL,
    code = 'ASL <- radsl(seed = 1); keys(ASL) <- c("USUBJID", "STUDYID")',
    check = FALSE
  ),
  root_modules(
    tm_table(
      "Table Choices",
      dataname =  "ASL",
      xvar = list(asl_extract_xvar),
      yvar = list(asl_extract_yvar)
    )
  )
)

#> |apply filter for ASL
#> |** UI id for modules is teal_modules.root
#> |UI id for module is teal_modules_root.Table_Choices

# NOT RUN {
shinyApp(app$ui, app$server)
# }