returns a data frame that stores the trend line statistics of a polynomial line fitted to scatterplot

returns a data frame that stores the trend line statistics of a polynomial line fitted to scatterplot

trend_line_stats(
  data_df,
  x_var,
  y_var,
  smoothing_degree,
  group_by_var = NULL,
  facet_by_var = NULL
)

Arguments

data_df
(data.frame) the data where the x-axis and y-axis is host

x_var
(character) the name of the column in `data_df` to be used as the x-axis

y_var
(character) the name of the column in `data_df` to be used as the y-axis

smoothing_degree
(integer) the degree of the polynomial line to be fitted

group_by_var
optional, (character) the name of the column in `data_df` that is used for grouping. This parameter is only used for determining whether a row in the output will have multiple groups or only one group.

facet_by_var
optional, (character) a vector of the name(s) of the column(s) in `data_df` to be used for filtering. Each row of the output will consists of trend line statistics for a single combination of all the possible combinations of the matrix created by all of the unique values of columns in this parameter.

Examples

if (FALSE) {
trend_line_stats(mpg, "displ", "hwy", 2)
trend_line_stats(mpg, "displ", "hwy", 2, group_by_var = "class", facet_by_var = "manufacturer")
}