Plot variable

Creates summary plot with statistics relevant to data type.

plot_var_summary(
  var,
  var_lab,
  numeric_as_factor,
  display_density = is.numeric(var),
  outlier_definition,
  records_for_factor
)

Arguments

var
vector of any type to be plotted. For numeric variables it produces histogram with density line, for factors it creates frequency plot

var_lab
text describing selected variable to be displayed on the plot

numeric_as_factor
logical should the numeric variable be treated as a factor

display_density
logical Should density estimation be displayed for numeric values?

outlier_definition
If 0 no outliers are removed, otherwise outliers (those more than outlier_definition*IQR below/above Q1/Q3 be removed)

records_for_factor
numeric if the number of factor levels is >= than this value then a graph of the factors isn't shown, only a list of values.

Value

plot