Scatterplot and Regression Model

Scatterplot and Regression Model

tm_a_regression(
  label = "Regression Analysis",
  regressor,
  response,
  plot_height = c(600, 200, 2000),
  plot_width = NULL,
  alpha = c(1, 0, 1),
  size = c(2, 1, 8),
  ggtheme = gg_themes,
  ggplot2_args = teal.devel::ggplot2_args(),
  pre_output = NULL,
  post_output = NULL,
  default_plot_type = 1,
  default_outlier_label = "USUBJID"
)

Arguments

label

(character) Label shown in the navigation item for the module.

regressor

(data_extract_spec or list of multiple data_extract_spec) Regressor variables from an incoming dataset with filtering and selecting.

response

(data_extract_spec or list of multiple data_extract_spec) Response variables from an incoming dataset with filtering and selecting.

plot_height

optional, (numeric) A vector of length three with c(value, min and max) for a slider encoding the plot height.

plot_width

optional, (numeric) A vector of length three with c(value, min and max) for a slider encoding the plot width.

alpha

optional, (numeric) If scalar then the plot points will have a fixed opacity. If a slider should be presented to adjust the plot point opacity dynamically then it can be a vector of length three with c(value, min, max).

size

optional, (numeric) If scalar then the plot point sizes will have a fixed size If a slider should be presented to adjust the plot point sizes dynamically then it can be a vector of length three with c(value, min, max).

ggtheme

optional, (character) ggplot Theme to be used by default. gg_themes is defined internally as c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic", "void", "test") All themes can be chosen by the user. Defaults to gray.

ggplot2_args

optional, (`ggplot2_args`) object created by [`teal.devel::ggplot2_args()`] with settings for all the plots or named list of `ggplot2_args` objects for plot-specific settings. For more details see the help vignette: `vignette("Custom ggplot2 arguments module", package = "teal.devel")`. List names should match the following:
`c("default", "Response vs Regressor", "Residuals vs Fitted", "Scale-Location", "Cook's distance", "Residuals vs Leverage"", "Cook's dist vs Leverage")`. The argument is merged with options variable `teal.ggplot2_args` and default module setup.

pre_output

optional, shiny.tag with text placed before the output to put the output into context. For example a title.

post_output

optional, shiny.tag with text placed after the output to put the output into context. For example the helpText elements are useful.

default_plot_type

optional, (numeric) Defaults to Response vs Regressor.

Response vs Regressor

Residuals vs Fitted

Normal Q-Q

Scale-Location

Cook's distance

Residuals vs Leverage

Cook's dist vs Leverage

default_outlier_label

optional, (character) The default column selected to label outliers.

Note

For more examples, please see the vignette "Using regression plots" via vignette("using-regression-plots", package = "teal.modules.general").

Examples

# Regression graphs from selected response variable (BMRKR1) and
# selected regressors (AGE)
library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"),
    check = TRUE
  ),
  modules = root_modules(
    tm_a_regression(
      label = "Regression",
      response = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = "BMRKR1",
          selected = "BMRKR1",
          multiple = FALSE,
          fixed = TRUE
        )
      ),
      regressor = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variables:",
          choices = variable_choices(ADSL, c("AGE", "SEX", "RACE")),
          selected = "AGE",
          multiple = TRUE,
          fixed = FALSE
        )
      )
    )
  )
)
#> [INFO] 2022-01-13 14:36:57.7932 pid:1855 token: teal.modules.general Initializing tm_a_regression
if (FALSE) {
shinyApp(app$ui, app$server)
}