Data Table Viewer Teal Module

A data table viewer shows the data using a paginated table.

tm_data_table(
  label = "Data table",
  variables_selected = list(),
  datasets_selected = character(0)
)

Arguments

label
(character) Label shown in the navigation item for the module.

variables_selected
(list) A named list of variables which should be initially shown for particular dataset. Names in list should correspond with names provided in list data(). If not specified for any dataset, the first six variables from the dataset will be shown.

datasets_selected
(character) A vector of datasets which should be shown and in what order. Names in the vector have to correspond with datasets names.

Examples

library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = "ADSL <- radsl(cached = TRUE)",
    check = TRUE
  ),
  modules = root_modules(
    tm_data_table(
      variables_selected = list(ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"))
    )
  )
)
if (FALSE) {
shinyApp(app$ui, app$server)
}

# two-datasets example
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADTTE <- radaette(cached = TRUE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADTTE", ADTTE),
    code = "ADSL <- radsl(cached = TRUE); ADTTE <- radaette(cached = TRUE)",
    check = TRUE
  ),
  modules = root_modules(
    tm_data_table(
      variables_selected = list(
        ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"),
        ADTTE = c(
          "STUDYID", "USUBJID", "SUBJID", "SITEID",
          "PARAM", "PARAMCD", "ARM", "ARMCD", "AVAL", "CNSR"
        )
      )
    )
  )
)
if (FALSE) {
shinyApp(app$ui, app$server)
}

# datasets: different subsets of long dataset
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADLB <- radlb(cached = TRUE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADLB", ADLB),
    code = "ADSL <- radsl(cached = TRUE); ADLB <- radlb(cached = TRUE)",
    check = TRUE
  ),
  modules = root_modules(
    tm_data_table(
      variables_selected = list(
        ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"),
        ADLB = c(
          "STUDYID", "USUBJID", "SUBJID", "SITEID",
          "PARAM", "PARAMCD", "AVISIT", "AVISITN", "AVAL", "CHG"
        )
      )
    )
  )
)
if (FALSE) {
shinyApp(app$ui, app$server)
}
# datasets: subsetting or changing order of datasets inside tm_data_table
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADLB <- radlb(cached = TRUE)
ADTTE <- radtte(cached = TRUE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADLB", ADLB),
    cdisc_dataset("ADTTE", ADTTE),
    code = "ADSL <- radsl(cached = TRUE);
            ADLB <- radlb(cached = TRUE);
            ADTTE <- radtte(cached = TRUE);",
    check = TRUE
  ),
  modules = root_modules(
    tm_data_table(
      variables_selected = list(
        ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"),
        ADLB = c(
          "STUDYID", "USUBJID", "SUBJID", "SITEID",
          "PARAM", "PARAMCD", "AVISIT", "AVISITN", "AVAL", "CHG"
        )
      ),
      datasets_selected = c("ADLB", "ADSL")
    )
  )
)
if (FALSE) {
shinyApp(app$ui, app$server)
}