Stack Plots of variables and show association with reference variable

Stack Plots of variables and show association with reference variable

tm_g_association(
  label = "Association",
  ref,
  vars,
  show_association = TRUE,
  plot_height = c(600, 400, 5000),
  plot_width = NULL,
  distribution_theme = gg_themes,
  association_theme = gg_themes,
  pre_output = NULL,
  post_output = NULL
)

Arguments

label
(character) Label shown in the navigation item for the module.

ref
(data_extract_spec or list of multiple data_extract_spec) reference variable, must set multiple = FALSE.

vars
(data_extract_spec or list of multiple data_extract_spec) associated variables.

show_association
optional, (logical) Whether show association of vars with reference variable. Defaults to TRUE.

plot_height
optional, (numeric) A vector of length three with c(value, min and max) for a slider encoding the plot height.

plot_width
optional, (numeric) A vector of length three with c(value, min and max) for a slider encoding the plot width.

distribution_theme
optional, (character) ggplot Theme to be used by default for the distribution plot. All themes can be chosen by the user. Defaults to gray. gg_themes is defined internally as c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic", "void", "test")

association_theme
optional, (character) ggplot Theme to be used by default for the association plots. All themes can be chosen by the user. Defaults to gray. gg_themes is defined internally as c("gray", "bw", "linedraw", "light", "dark", "minimal", "classic", "void", "test")

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example a title.

post_output
optional, shiny.tag with text placed after the output to put the output into context. For example the helpText elements are useful.

Note

For more examples, please see the vignette "Using association plot" via vignette("using-association-plot", package = "teal.modules.general").

Examples

# Association plot of selected reference variable (SEX)
# against other selected variables (BMRKR1)
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- radsl(cached = TRUE)"),
    check = TRUE
  ),
  modules = root_modules(
    tm_g_association(
      ref = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variable:",
          choices = variable_choices(
            ADSL,
            c("SEX", "RACE", "COUNTRY", "ARM", "STRATA1", "STRATA2", "ITTFL", "BMRKR2")
          ),
          selected = "RACE",
          fixed = FALSE
        )
      ),
      vars = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          label = "Select variables:",
          choices = variable_choices(
            ADSL,
            c("SEX", "RACE", "COUNTRY", "ARM", "STRATA1", "STRATA2", "ITTFL", "BMRKR2")
          ),
          selected = "BMRKR2",
          multiple = TRUE,
          fixed = FALSE
        )
      )
    )
  )
)
if (FALSE) {
shinyApp(app$ui, app$server)
}