Stack Plots of variables and show association with reference variable

Stack Plots of variables and show association with reference variable

tm_g_association(label = "Association", dataname, var,
  show_association = TRUE, plot_height = c(600, 400, 5000),
  pre_output = NULL, post_output = NULL, with_show_r_code = TRUE)

Arguments

label
label shown in the navigation for the item

dataname
(character) data set name to analyze

var
(choices_selected) TODO

show_association
(logical) TODO

plot_height
(numeric) vector with three elements defining selected, min and max plot height

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

with_show_r_code
(logical) Whether show R Code button shall be enabled

Examples

library(random.cdisc.data)
ASL <- radsl(seed = 1)

app <- init(
  data = cdisc_data(ASL = ASL),
  modules = root_modules(
    tm_g_association(
      dataname = "ASL",
      var = choices_selected(names(ASL), "AGE")
    )
  )
)

#> |apply filter for ASL
#> |** UI id for modules is teal_modules.root
#> |UI id for module is teal_modules_root.Association

# NOT RUN {
shinyApp(app$ui, app$server)
# }